Refactoring
- Pulls out commonly used logic to common/
- Add parent POM
