+++ aws-nimblestudio-launchprofile/template.yml
@@ -13,11 +13,11 @@ Resources:
     Properties:
       Handler: software.amazon.nimblestudio.launchprofile.HandlerWrapper::handleRequest
       Runtime: java8
-      CodeUri: ./target/aws-nimblestudio-launchprofile-1.0.jar
+      CodeUri: ./target/aws-nimblestudio-launchprofile-handler-1.0-SNAPSHOT.jar
 
   TestEntrypoint:
     Type: AWS::Serverless::Function
     Properties:
       Handler: software.amazon.nimblestudio.launchprofile.HandlerWrapper::testEntrypoint
       Runtime: java8
-      CodeUri: ./target/aws-nimblestudio-launchprofile-1.0.jar
+      CodeUri: ./target/aws-nimblestudio-launchprofile-handler-1.0-SNAPSHOT.jar
