+++ aws-nimblestudio-studio/template.yml
@@ -13,11 +13,11 @@ Resources:
     Properties:
       Handler: software.amazon.nimblestudio.studio.HandlerWrapper::handleRequest
       Runtime: java8
-      CodeUri: ./target/aws-nimblestudio-studio-1.0.jar
+      CodeUri: ./target/aws-nimblestudio-studio-handler-1.0-SNAPSHOT.jar
 
   TestEntrypoint:
     Type: AWS::Serverless::Function
     Properties:
       Handler: software.amazon.nimblestudio.studio.HandlerWrapper::testEntrypoint
       Runtime: java8
-      CodeUri: ./target/aws-nimblestudio-studio-1.0.jar
+      CodeUri: ./target/aws-nimblestudio-studio-handler-1.0-SNAPSHOT.jar
