+++ aws-nimblestudio-studiocomponent/template.yml
@@ -13,11 +13,11 @@ Resources:
     Properties:
       Handler: software.amazon.nimblestudio.studiocomponent.HandlerWrapper::handleRequest
       Runtime: java8
-      CodeUri: ./target/aws-nimblestudio-studiocomponent-1.0.jar
+      CodeUri: ./target/aws-nimblestudio-studiocomponent-handler-1.0-SNAPSHOT.jar
 
   TestEntrypoint:
     Type: AWS::Serverless::Function
     Properties:
       Handler: software.amazon.nimblestudio.studiocomponent.HandlerWrapper::testEntrypoint
       Runtime: java8
-      CodeUri: ./target/aws-nimblestudio-studiocomponent-1.0.jar
+      CodeUri: ./target/aws-nimblestudio-studiocomponent-handler-1.0-SNAPSHOT.jar
