+++ buildspec.yml
@@ -17,6 +17,9 @@ phases:
           # skip hidden folders
           dirs=$(find . -not -path "\./\.*" -mindepth 1 -maxdepth 1 -type d)
           echo "Folders to build: $dirs"
+          cd common
+          mvn -Dorg.slf4j.simpleLogger.log.org.apache.maven.cli.transfer.Slf4jMavenTransferListener=warn -B clean install --no-transfer-progress
+          cd ..
           for directory in $dirs; do
             cd "$directory"
             mvn -Dorg.slf4j.simpleLogger.log.org.apache.maven.cli.transfer.Slf4jMavenTransferListener=warn -B clean verify --no-transfer-progress
