+++ common/.gitignore
@@ -1,0 +1,25 @@
+# macOS
+.DS_Store
+._*
+
+# Maven outputs
+.classpath
+
+# IntelliJ
+*.iml
+.idea
+out.java
+out/
+.settings
+.project
+
+# auto-generated files
+target/
+build/
+docs/
+
+# our logs
+rpdk.log*
+
+# contains credentials
+sam-tests/
