+++ common/README.md
@@ -1,0 +1,2 @@
+## Amazon Nimble Studio Common
+This package contains logic that is common to all of Nimble Studio's CloudFormation resource types.
