+++ common/src/main/java/software/amazon/nimblestudio/common/ClientBuilder.java
@@ -1,4 +1,4 @@
-package software.amazon.nimblestudio.studiocomponent;
+package software.amazon.nimblestudio.common;
 
 import software.amazon.awssdk.core.client.config.ClientOverrideConfiguration;
 import software.amazon.awssdk.core.retry.RetryPolicy;
@@ -7,15 +7,12 @@ import software.amazon.awssdk.services.nimble.NimbleClient;
 import software.amazon.cloudformation.LambdaWrapper;
 
 public class ClientBuilder {
-
-    private static final int NUM_RETRY = 6;
-
     private ClientBuilder() {
     }
 
     private static final RetryPolicy RETRY_POLICY = RetryPolicy
         .builder()
-        .numRetries(NUM_RETRY)
+        .numRetries(6)
         .retryCondition(RetryCondition.defaultRetryCondition())
         .build();
